Refactor: 커뮤니티 페이지 게시글 목록 함수 모듈화
