+++ pages/community/qna/index.html
@@ -7,9 +7,9 @@
     <link rel="stylesheet" href="/styles/base-styles.css" />
     <link rel="stylesheet" href="/components/header/header.css" />
     <link rel="stylesheet" href="/components/footer/footer.css" />
-    <script type="module" src="/components/component.js"></script>
     <link rel="stylesheet" href="style.css" />
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
+    <script type="module" src="/components/component.js"></script>
     <title>Document</title>
   </head>
   <body>
